--- constants/meta/ClickOnaction-meta-data/clickOn-meta.ts
@@ -1,8 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const clickOnMeta: Meta = {
-    title: "ClickOn Action",
-    description:
-        "Defines a click action that can target elements either by CSS selector or by visible text content.",
-};
