--- constants/meta/Scroll-meta-data/scroll-direction-meta.ts
@@ -1,7 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const scrollDirectionMeta: Meta = {
-    description: "Direction to scroll the webpage.",
-    enumDescriptions: ["Scrolls the page upwards", "Scrolls the page downwards"],
-};
