--- constants/meta/Scroll-meta-data/scroll-meta.ts
@@ -1,6 +1,0 @@
-import { Meta } from "../json-schema-meta.ts";
-
-export const scrollMeta: Meta = {
-    title: "Scroll Action",
-    description: "Defines a horizontal or vertical scroll action on the webpage.",
-};
