--- constants/meta/Scroll-meta-data/scroll-posY-meta.ts
@@ -1,5 +1,0 @@
-import { Meta } from "../json-schema-meta.ts";
-
-export const scrollPositionYMeta: Meta = {
-    description: "Vertical position to scroll to on the webpage.",
-};
