--- constants/meta/Scroll-meta-data/scroll-postionX-meta.ts
@@ -1,6 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const scrollPositionXMeta: Meta = {
-    description: "Horizontal position to scroll to on the webpage.",
-};
