--- constants/meta/core/flow-meta.ts
@@ -1,9 +1,0 @@
-
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const flowMeta: Meta = {
-	title: "Workflow Flow Steps",
-	description:
-		"Ordered sequence of browser automation steps (clicks, inputs, scrolling, keyboard actions, waits, etc.) that define the complete workflow execution path.",
-};
