--- constants/meta/core/name-meta.ts
@@ -1,8 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const nameMeta: Meta = {
-	title: "Workflow Name",
-	description:
-		"Name of your workflow. Symphony will display the name when running the workflow.",
-};
