--- constants/meta/core/url-meta.ts
@@ -1,8 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const urlMeta: Meta = {
-	title: "Target Website URL",
-	description:
-		"Target website URL where this workflow will be executed. This determines the starting page for all automation steps in the flow.",
-};
