+++ constants/meta/flow/input/index-input.ts
@@ -1,5 +1,17 @@
 
-import { Meta } from "../json-schema-meta.ts";
+import { Meta } from "../../json-schema-meta.ts";
+
+export const inputMeta: Meta = {
+    title: "Input Action",
+    description:
+        "Defines an input action where a value is entered into a specified input field identified by the `selector`, `label`, `testID`, `placeholder` or `role` .",
+};
+
+
+
+
+
+
 
 
 export const inputSelectorMeta: Meta = {
@@ -46,3 +58,11 @@ export const inputSelectorMeta: Meta = {
         },
     ],
 };
+
+
+
+
+export const inputValueMeta: Meta = {
+    description: "Value to be entered into the input field.",
+};
+
