+++ constants/meta/flow/keyboardAction-meta-data/index-keyboard.ts
@@ -1,6 +1,14 @@
 
+import { Meta } from "../../json-schema-meta.ts";
 
-import { Meta } from "../json-schema-meta.ts";
+export const keyboardMeta: Meta = {
+    title: "Keyboard Action",
+    description:
+        "Defines a keyboard action that simulates key presses on the webpage.",
+};
+
+
+
 
 export const keyboardKeyMeta: Meta = {
     markdownDescription:
