+++ constants/meta/index.ts
@@ -1,34 +1,6 @@
-export * from "../meta/ClickOnaction-meta-data/clickOn-String";
-export * from "../meta/ClickOnaction-meta-data/clickOn-meta";
-export * from "../meta/ClickOnaction-meta-data/clickOn-selector";
-
-export * from "../meta/core/color-meta";
-export * from "../meta/core/flow-meta";
-export * from "../meta/core/url-meta";
-export * from "../meta/core/name-meta";
-
-
-export * from "../meta/input/input-meta";
-
-export * from "../meta/input/input-selector-meta";
-
-export * from "../meta/input/input-value"
-
-
-
-
-export * from "../meta/keyboardAction-meta-data/keyboard-meta";
-export * from "../meta/keyboardAction-meta-data/keyboardkey-meta";
-
-export * from "../meta/Scroll-meta-data/scroll-direction-meta";
-export * from "../meta/Scroll-meta-data/scroll-meta";
-export * from "../meta/Scroll-meta-data/scroll-postion-meta";
-export * from "../meta/Scroll-meta-data/scroll-postionX-meta";
-export * from "../meta/Scroll-meta-data/scroll-posY-meta";
-export * from "../meta/Scroll-meta-data/scroll-speed-meta";
-
-
-
-export * from "../meta/waitforaction-meta-data/waitfor-duration"
-export * from "../meta/waitforaction-meta-data/waitfor-meta"
-
+export * from "./flow/ClickOnaction-meta-data/index-clickOn.ts";
+export * from "./flow/core/index-core.ts";
+export * from "./flow/keyboardAction-meta-data/index-keyboard.ts";
+export * from "./flow/Scroll-meta-data/index-scroll.ts";
+export * from "./flow/waitforaction-meta-data/index-waitfor.ts";    
+export * from "./flow/input/index-input.ts";
