--- constants/meta/input/input-value.ts
@@ -1,6 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const inputValueMeta: Meta = {
-    description: "Value to be entered into the input field.",
-};
