--- constants/meta/keyboardAction-meta-data/keyboard-meta.ts
@@ -1,8 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const keyboardMeta: Meta = {
-    title: "Keyboard Action",
-    description:
-        "Defines a keyboard action that simulates key presses on the webpage.",
-};
