--- constants/meta/waitforaction-meta-data/waitfor-meta.ts
@@ -1,8 +1,0 @@
-
-import { Meta } from "../json-schema-meta.ts";
-
-export const waitForMeta: Meta = {
-    title: "WaitFor Action",
-    description:
-        "Defines a wait action that pauses workflow execution for a specified duration in milliseconds.",
-};
